+++ server/server/middlewares/auth.middleware.ts
@@ -33,8 +33,9 @@ export const authMiddleware = async (
       return next();
     }
     const access_token = req.cookies[process.env.SUPABASE_COOKIE_KEY!];
-    const { data: user, error } =
-      await supabaseClient.auth.getUser(access_token);
+    const { data: user, error } = await supabaseClient.auth.getUser(
+      access_token,
+    );
 
     if (error || !user) {
       log.error(error);
