+++ server/server/middlewares/proxy.middleware.ts
@@ -53,8 +53,9 @@ export class ProxyMiddleware {
   async chooseRoute(req: Request) {
     const host = req.hostname;
     const subdomain = host.split(".")[0];
-    const selectedSniffer =
-      await this.snifferService.findBySubdomain(subdomain);
+    const selectedSniffer = await this.snifferService.findBySubdomain(
+      subdomain,
+    );
     req.headers["x-sharkio-port"] = selectedSniffer?.port.toString();
 
     if (selectedSniffer != null) {
